+++ example/src/layouts/components/Container/index.jsx
@@ -1,12 +1,12 @@
-import React, { Component } from "react"
+import React, { Component, useEffect } from "react"
 import { NavLink, connect, useConnect, useNuomi, withNuomi, router, ShapeRoute } from "nuomi"
 
 export default ({ children }) => {
-  const [{ count, routes }, dispatch] = useConnect(({ routes, count }) => ({ routes, count }));
+  const [{ count, routes }, dispatch] = useConnect();
   const nuomi = useNuomi();
   return (
     <div>
-      <NavLink to={{ path: '/', state: { a: 1 } }}>首页</NavLink>
+      <NavLink to={{ path: '/' }}>首页</NavLink>
       <NavLink to={{ path: '/list' }} style={{ marginLeft: 10 }}>列表</NavLink>
 
       <div onClick={() => {