+++ packages/nuomi/src/core/redux/reducer.js
@@ -1,12 +1,12 @@
 import { combineReducers } from 'redux';
-import globalStore, { setStore, getStore } from './store';
+import globalStore, { setStore } from './store';
 import { noop as rootReducer } from '../../utils';
-import globalWindow from '../../utils/globalWindow';
 
 const reducers = {};
 
 function replaceReducer() {
-  if (Object.keys(reducers).length) {
+  const keys = Object.keys(reducers);
+  if (keys.length) {
     globalStore.replaceReducer(combineReducers(reducers));
   } else {
     globalStore.replaceReducer(rootReducer);
@@ -22,13 +22,6 @@ export function createReducer(key, reducer) {
 
 export function removeReducer(key) {
   if (reducers[key]) {
-    // 解决REDUX_DEVTOOLS缓存状态问题
-    if (process.env.NODE_ENV !== 'production' && globalWindow.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) {
-      const store = getStore(key);
-      if (store) {
-        store.restoreState();
-      }
-    }
     delete reducers[key];
     setStore(key, null);
     replaceReducer();